+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Change Log
 - add `SKTileObject.tileData` property
 - add `SKTiledSceneCamera.clampZoomValue`
 - add `SKTiledSceneCamera.zoomClamping` property
+- add `SKTiledLayerObject.isStatic` property
 - remove `SKTile.pauseAnimation`